# static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "v1",
        "title": "Bordo-Estate API",
        "description": "Estate API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [{
            "url": "https://estate-bordo-55.herokuapp.com/",
            "description": "heroku"
        },
        {
            "url": "https://5498-88-248-198-29.eu.ngrok.io/",
            "description": "ngrok"
        },
        {
            "url": " http://localhost:5000/",
            "description": "local"
        }
    ],
...
    "paths": {
...
        "/categories": {
            "summary": "",
            "description": "",
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Category get",
                "description": "Category sub category",
                "operationId": "users.category",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [{
                                            "$ref": "#/components/schemas/login/allOf/0"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "allOf": [{
                                                                "type": "object",
                                                                "properties": {
                                                                    "_id": {
                                                                        "title": "Document MongoDb Id",
                                                                        "description": "MongoDb Id of the document",
                                                                        "type": "string",
                                                                        "example": "62ed69679c045c97a63390a4"
                                                                    },
                                                                    "categoryName": {
                                                                        "title": "Category Name",
                                                                        "description": "Category Name",
                                                                        "type": "string",
                                                                        "example": "string"
                                                                    },
                                                                    "parent": {
                                                                        "title": "Document MongoDb Id",
                                                                        "description": "MongoDb Id of the document",
                                                                        "type": "string",
                                                                        "example": "62ed69679c045c97a63390a4"
                                                                    },
                                                                    "path": {
                                                                        "title": "Category path by Id",
                                                                        "description": "Category path",
                                                                        "type": "string",
                                                                        "example": "62ed69679c045c97a63390a4,62ed69679c045c97a63390a4,62ed69679c045c97a63390a4"
                                                                    }
                                                                }
                                                            },
                                                            {
                                                                "type": "object",
                                                                "properties": {
                                                                    "children": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "$ref": "#/paths/~1categories/get/responses/200/content/application~1json/schema/allOf/1/properties/data/items/allOf/0"
                                                                        }
                                                                    }
                                                                }
                                                            }
                                                        ]
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/paths/~1adverts/get/responses/400"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}